turbo UI refactoring
