+++ VisualChatGPTStudio2019/VisualChatGPTStudio2019.csproj
@@ -150,6 +150,9 @@
     <PackageReference Include="Microsoft.Identity.Client.Extensions.Msal">
       <Version>4.71.1</Version>
     </PackageReference>
+    <PackageReference Include="Microsoft.VisualStudio.Imaging">
+      <Version>17.14.40264</Version>
+    </PackageReference>
     <PackageReference Include="Microsoft.VSSDK.BuildTools" Version="17.9.3184">
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
       <PrivateAssets>all</PrivateAssets>
