--- VisualChatGPTStudioShared/ToolWindows/Turbo/ucChatHeader.xaml
@@ -1,40 +1,0 @@
-﻿<UserControl x:Class="JeffPires.VisualChatGPTStudio.ToolWindows.Turbo.ucChatHeader"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-             xmlns:i="http://schemas.microsoft.com/expression/2010/interactivity" 
-             xmlns:vsshell="clr-namespace:Microsoft.VisualStudio.Shell;assembly=Microsoft.VisualStudio.Shell.15.0"   
-             xmlns:utils="clr-namespace:JeffPires.VisualChatGPTStudio.Utils"             
-             mc:Ignorable="d"             
-             Name="MyToolWindow">
-    <Grid>
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition Width="Auto" />
-            <ColumnDefinition Width="10" />
-            <ColumnDefinition Width="Auto" />
-        </Grid.ColumnDefinitions>
-        <TextBlock Grid.Column="0" Name="lblHeader" Foreground="{DynamicResource {x:Static vsshell:VsBrushes.WindowTextKey}}"/>
-        <Image 
-            Grid.Column="2" 
-            Height="10" 
-            Width="10" 
-            HorizontalAlignment="Left" 
-            Name="btnClose" 
-            MouseDown="btnClose_Click" 
-            ToolTip="Close Tab (CTRL+W)" 
-            AutomationProperties.HelpText="Close Tab (CTRL+W)"
-            Source="pack://application:,,,/VisualChatGPTStudio;component/Resources/close.png">
-            <Image.Style>
-                <Style TargetType="{x:Type Image}">
-                    <Setter Property="Cursor" Value="Arrow"/>
-                    <Style.Triggers>
-                        <Trigger Property="IsMouseOver" Value="True">
-                            <Setter Property="Cursor" Value="Hand"/>
-                        </Trigger>
-                    </Style.Triggers>
-                </Style>
-            </Image.Style>
-        </Image>
-    </Grid>  
-</UserControl>
